# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2026.5.16-1108mdt
  description: QuantumShield — Post-quantum encrypted chat API (ML-KEM-1024 + ML-DSA-87)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: getUsersSearch
      tags: [users]
      summary: Search active identity codes
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUsersUserId
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms:
    get:
      operationId: getRooms
      tags: [rooms]
      summary: List rooms for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"

    post:
      operationId: postRooms
      tags: [rooms]
      summary: Create a new room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    get:
      operationId: getRoomsRoomId
      tags: [rooms]
      summary: Get room details
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteRoomsRoomId
      tags: [rooms]
      summary: Delete a room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rooms/{roomId}/members:
    get:
      operationId: getRoomsRoomIdMembers
      tags: [rooms]
      summary: Get room members
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: postRoomsRoomIdMembers
      tags: [rooms]
      summary: Add a member to a room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "200":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
          description: Active identity code for compatibility with older clients; may be "sealed".
        primaryCode:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarColor:
          type: string
          nullable: true
        kemPublicKey:
          type: string
          nullable: true
          description: Base64-encoded ML-KEM-1024 public key
        dsaPublicKey:
          type: string
          nullable: true
          description: Base64-encoded ML-DSA-87 public key
        createdAt:
          type: string
          format: date-time
...
    Room:
      type: object
      required: [id, type, createdAt, memberCount]
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        type:
          type: string
          enum: [direct, group]
        memberCount:
          type: integer
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        ttlSeconds:
          type: integer
          nullable: true
          description: Message TTL — keys destroyed after this period
        ttlMode:
          type: string
          enum: [after_view, after_send]
          default: after_view
          description: Whether TTL starts when a message is first viewed or when it is sent.
        deliveryFuzzSeconds:
          type: integer
          default: 89
          description: Random server-side delivery delay window in seconds.
        decayMode:
          type: string
          enum: [standard, experimental_quorum_decay]
          default: standard
          description: Optional experimental quorum-decay policy tier. Standard is the default.
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/User"
    CreateRoomRequest:
      type: object
      required: [type, memberIds]
      properties:
        name:
          type: string
          nullable: true
        type:
          type: string
          enum: [direct, group]
        memberIds:
          type: array
          items:
            type: string
        ttlSeconds:
          type: integer
          nullable: true
          description: Message TTL in seconds. Keys are destroyed after this period.
        ttlMode:
          type: string
          enum: [after_view, after_send]
          default: after_view
          description: Whether TTL starts when a message is first viewed or when it is sent.
        deliveryFuzzSeconds:
          type: integer
          default: 89
          description: Random server-side delivery delay window in seconds.
        decayMode:
          type: string
          enum: [standard, experimental_quorum_decay]
          default: standard
          description: Optional experimental quorum-decay policy tier. Standard is the default.
...
    AddMemberRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
...