# lib/api-zod/src/generated/types/createRoomRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumShield — Post-quantum encrypted chat API (ML-KEM-1024 + ML-DSA-87)
 * OpenAPI spec version: 2026.5.16-1108mdt
 */
import type { CreateRoomRequestTtlMode } from './createRoomRequestTtlMode';
import type { CreateRoomRequestType } from './createRoomRequestType';
import type { CreateRoomRequestDecayMode } from './createRoomRequestDecayMode';

export interface CreateRoomRequest {
  name?: string | null;
  type: CreateRoomRequestType;
  memberIds: string[];
  /** Message TTL in seconds. Keys are destroyed after this period. */
  ttlSeconds?: number | null;
  /** Whether TTL starts when a message is first viewed or when it is sent. */
  ttlMode?: CreateRoomRequestTtlMode;
  /** Random server-side delivery delay window in seconds. */
  deliveryFuzzSeconds?: number;
  /** Optional experimental quorum-decay policy tier. Standard is the default. */
  decayMode?: CreateRoomRequestDecayMode;
}