# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumShield — Post-quantum encrypted chat API (ML-KEM-1024 + ML-DSA-87)
 * OpenAPI spec version: 2026.5.16-1108mdt
 */
import type { RoomTtlMode } from './roomTtlMode';
import type { RoomType } from './roomType';
import type { RoomDecayMode } from './roomDecayMode';
import type { User } from './user';

export interface Room {
  id: string;
  name?: string | null;
  type: RoomType;
  memberCount: number;
  lastMessageAt?: Date | null;
  /** Message TTL — keys destroyed after this period */
  ttlSeconds?: number | null;
  /** Whether TTL starts when a message is first viewed or when it is sent. */
  ttlMode?: RoomTtlMode;
  /** Random server-side delivery delay window in seconds. */
  deliveryFuzzSeconds?: number;
  /** Optional experimental quorum-decay policy tier. Standard is the default. */
  decayMode?: RoomDecayMode;
  createdAt: Date;
  members?: User[] | null;
}